multi-threaded added
I use a multi-threaded approach to speed up the password-cracking process. The crack_passwords function is run in a separate thread for each chunk of passwords read from the file. Each thread calls the crack method to try the passwords in the chunk.

Note that this code is still a brute-force approach, which can be very slow if the password list is huge.